requirements.txt
pillow
requests
spacy
scikit-learn
numpy
